requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.44
pydantic>=2.12.5
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
googlemaps>=4.2.0
requests>=2.32.3
Jinja2>=3.1.4
python-multipart>=0.0.20
bcrypt>=3.2.2
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.11
pyasn1>=0.4.6,<0.5.0
argon2-cffi>=23.1.0slowapi>=0.1.9
python-multipart>=0.0.20